requirements.txt
gradio
Pillow
requests
flask
flask_cors
transformers